+++ Cargo.toml
@@ -18,18 +18,17 @@ categories = ["multimedia","os", "wasm"]
 [features]
 clipboard = ["dep:clipboard"]
 notifications = ["dep:notify-rust"]
-geolocation = ["dep:windows", "dep:futures-util", "dep:futures", "dep:web-sys", "dep:wasm-bindgen", "rw"]
-channel = ["dep:async-broadcast", "dep:uuid"]
+geolocation = ["dep:windows", "dep:futures-util", "dep:futures", "dep:web-sys", "dep:wasm-bindgen", "utils"]
 color_scheme = ["dep:web-sys", "dep:wasm-bindgen", "dep:wasm-bindgen-futures"]
-rw = []
+utils = ["dep:async-broadcast", "dep:uuid"]
 i18n = ["dep:serde", "dep:serde_json", "dep:unic-langid"]
-all = ["clipboard", "notifications", "geolocation", "channel", "color_scheme", "rw", "i18n"]
+all = ["clipboard", "notifications", "geolocation", "color_scheme", "utils", "i18n"]
 
 [dependencies]
 dioxus = { version = "0.3" }
 cfg-if = "1.0.0"
 
-# channel
+# utils
 uuid = { version = "1.3.2", features = ["v4"], optional = true }
 async-broadcast = { version = "0.5.1", optional = true }
 # clipboard
