+++ examples/README.md
@@ -5,3 +5,9 @@ Learn how to use `use_preferred_color_scheme`.
 
 ### [`geolocation`](./geolocation/)
 Learn how to use the `geolocation` abstraction.
+
+### [`i18n`](./i18n/)
+Learn how to use the `i18n` abstraction.
+
+### [`channel`](./channel/)
+Learn how to use the `channel` abstraction.
