+++ examples/channel/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-dioxus-std = { path="../../", features = ["channel"] }
+dioxus-std = { path="../../", features = ["utils"] }
 dioxus = "0.3"
 dioxus-web = "0.3"
