+++ examples/geolocation/README.md
@@ -4,4 +4,4 @@ Learn how to use the `geolocation` abstraction.
 
 Run:
 
-```dioxus serve```
+```cargo run```
