--- src/channel/mod.rs
@@ -1,2 +1,0 @@
-mod use_channel;
-pub use use_channel::*;
