+++ src/geolocation/mod.rs
@@ -1,2 +1,9 @@
-pub mod use_geolocation;
-pub mod geolocation;
+mod core;
+mod platform;
+mod use_geolocation;
+
+pub mod geolocation {
+    pub use super::core::*;
+    pub use super::platform::*;
+    pub use super::use_geolocation::*;
+}
