+++ src/geolocation/platform/mod.rs
@@ -1,7 +1,7 @@
 cfg_if::cfg_if! {
     if #[cfg(windows)] {
         mod windows;
-        pub use self::windows::*;
+        pub use windows::*;
     } else if #[cfg(target_family = "wasm")] {
         mod wasm;
         pub use wasm::*;
