+++ src/geolocation/platform/windows.rs
