--- src/sys/windows/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "geolocation")]
-pub mod geolocation;
